+++ Demo.hs
@@ -8,24 +8,25 @@ import qualified Data.ByteString.Char8 as B
 import Data.Maybe (catMaybes, isJust)
 
 main =
-  do demonstrate "math"      math      "1 + 3"
-     demonstrate "math"      math      "3 * 2 + 100"
-     demonstrate "math"      math      "20 / 3 + 100 * 2"
-     demonstrate "peano"     peano     "S Z"
-     demonstrate "peano"     peano     "S S S S Z"
-     demonstrate "peano"     peano     "S   S   Z"
-     demonstrate "parsePair" parsePair "<-1, 3>"
-     demonstrate "parsePair" parsePair "<-4,3b0>"
-     demonstrate "parsePair" parsePair "< a,  -30 >"
-     demonstrate "parsePair" parsePair "< a,  other>"
-     demonstrate "parseDate" parseDate "1993.8.10"
-     demonstrate "parseDate" parseDate "1993.08.10"
-     demonstrate "parseDate" parseDate "2003.02.28"
-     demonstrate "parseDate" parseDate "2004.02.28"
-     demonstrate "parseDate" parseDate "2003.02.27"
-     demonstrate "disjunct"  disjunct  "a"
-     demonstrate "disjunct"  disjunct  "ab"
-     demonstrate "disjunct"  disjunct  "abc"
+  do demonstrate "math"       math       "1 + 3"
+     demonstrate "math"       math       "3 * 2 + 100"
+     demonstrate "math"       math       "20 / 3 + 100 * 2"
+     demonstrate "peano"      peano      "S Z"
+     demonstrate "peano"      peano      "S S S S Z"
+     demonstrate "peano"      peano      "S   S   Z"
+     demonstrate "parsePair"  parsePair  "<-1, 3>"
+     demonstrate "parsePair"  parsePair  "<-4,3b0>"
+     demonstrate "parsePair"  parsePair  "< a,  -30 >"
+     demonstrate "parsePair"  parsePair  "< a,  other>"
+     demonstrate "parsePair'" parsePair' "<-1, 3>"
+     demonstrate "parseDate"  parseDate  "1993.8.10"
+     demonstrate "parseDate"  parseDate  "1993.08.10"
+     demonstrate "parseDate"  parseDate  "2003.02.28"
+     demonstrate "parseDate"  parseDate  "2004.02.28"
+     demonstrate "parseDate"  parseDate  "2003.02.27"
+     demonstrate "disjunct"   disjunct   "a"
+     demonstrate "disjunct"   disjunct   "ab"
+     demonstrate "disjunct"   disjunct   "abc"
      print $ "btest: " ++ show btest
 
 demonstrate n f input = putStrLn $ n ++ " \"" ++ input ++ "\" == " ++ show (f input)
@@ -45,6 +46,11 @@ peano = [rex|^(?{ length . filter (=='S') } \s* (?:S\s+)*Z)\s*$|]
 
 parsePair :: String -> Maybe (String, String)
 parsePair = [rex|^<\s* (?{ }[^\s,>]+) \s*,\s* (?{ }[^\s,>]+) \s*>$|]
+
+parsePair' :: String -> Maybe (Int, Int)
+parsePair' = [rex|^<\s* (?{ }[^\s,>]+) \s*,\s* (?{ }[^\s,>]+) \s*>$|]
+  where
+    rexView = read
 
 -- From http://www.regular-expressions.info/dates.html
 parseDate :: String -> Maybe (Int, Int, Int)