+++ CHANGES.md
@@ -20,6 +20,17 @@
   * Made configuration into a datatype.
 
 0.4.2: Feb 4 2013
-------------------
+-----------------
   * Precompilation bugs fixed by [takano-akio](https://github.com/takano-akio)!
-    Now precompilation is on by default.
+
+0.4.3: Dec 21 2013
+------------------
+  * Patch from [aavogt](https://github.com/aavogt)! to use haskell-src-exts to
+    parse view patterns.
+  * Miscellaneous code cleanups.
+
+0.5: Feb 20 2014
+----------------
+  * Changed the configuration datatype to allow custom preprocessing of expr /
+    pat antiquotes.  The default match processing is now "rexView", an
+    identifier that can be shadowed locally or hidden on import.